fix: refine telemetry consent experience
