+++ packages/web/src/App.tsx
@@ -25,7 +25,7 @@ import { useViewportHeightCssVar } from './hooks/useViewportHeight';
 import { GlobalShortcutBridge } from './components/app/GlobalShortcutBridge';
 import { UpdateNotificationProvider } from './components/app/UpdateNotificationProvider';
 import { BootstrapStatusChip } from './components/app/BootstrapStatusChip';
-import { ObservabilityConsentModal } from './components/app/ObservabilityConsentModal';
+import { ObservabilityConsentBanner } from './components/app/ObservabilityConsentBanner';
 import { useLanguage } from './hooks/useLanguage';
 import { shouldShowServerStatusPage } from './store/radio/connectionView';
 
@@ -62,7 +62,7 @@ function AppContent() {
   const isCWMode = engineMode === 'cw';
 
   return (
-    <div className="App app-viewport-height w-full overflow-hidden relative">
+    <div className="App app-viewport-height relative flex w-full flex-col overflow-hidden">
       {/* PTT发射状态全局红色内描边 */}
       {pttStatus.isTransmitting && (
         <div
@@ -91,7 +91,10 @@ function AppContent() {
         defaultExtraWidth={26}
         minExtraWidth={18}
         maxExtraWidth={38}
+        className="!h-auto min-h-0 flex-1"
       />
+
+      <ObservabilityConsentBanner enabled={isAdmin && !showSetupOverlay} />
 
       <BootstrapStatusChip />
 
@@ -107,7 +110,6 @@ function AppContent() {
       {/* 首次使用引导（Admin）/ 观看者欢迎蒙层 */}
       <ProfileSetupOverlay isOpen={showSetupOverlay} />
       <ViewerWelcomeOverlay isOpen={showViewerWelcome} />
-      <ObservabilityConsentModal enabled={isAdmin && !showSetupOverlay} />
 
       {/* OpenWebRX SDR Profile 手动选择弹窗 */}
       <OpenWebRXProfileSelectModal />
