+++ packages/web/src/i18n/locales/en/common.json
@@ -256,22 +256,20 @@
   "system": {
     "title": "System Settings",
     "description": "Configure system-level feature options.",
-    "anonymousTelemetryTitle": "Anonymous usage statistics",
-    "anonymousTelemetryDesc": "With your consent, help us understand participating installations, usage, active versions, and concurrently online instances.",
-    "anonymousTelemetryConsent": "You have not made a privacy choice yet. No usage statistics are sent before you confirm it.",
+    "anonymousTelemetryConsent": "You have not chosen whether to participate. No statistics are sent until you do.",
     "anonymousTelemetryUnavailable": "The statistics endpoint is not configured in this build, so no data can be sent.",
     "anonymousTelemetryEnabled": "Enabled",
     "anonymousTelemetryDisabled": "Disabled",
-    "anonymousTelemetryConsentTitle": "Anonymous usage statistics and privacy",
-    "anonymousTelemetryConsentIntro": "TX-5DR would like to collect a minimal set of anonymous runtime data to understand real-world usage and improve future releases.",
-    "anonymousTelemetryCollectedTitle": "Collected",
-    "anonymousTelemetryCollected": "A randomly generated installation identifier; app version and build identifier, release channel, runtime type, operating-system family, and CPU architecture; process start, stop, and online state, uptime, and completed WebSocket connection count. The service stores only a one-way key derived from the installation identifier.",
-    "anonymousTelemetryNeverCollectedTitle": "Never collected",
-    "anonymousTelemetryNeverCollected": "Callsign, grid locator, QSO records, frequency, audio, radio model, device serial number, host name, user name, file paths, URLs, tokens, raw IP address, User-Agent, or log content. Diagnostic logs are never uploaded automatically.",
-    "anonymousTelemetryMeasurementNote": "Installation totals count participating installations; active and online totals count running instances; WebSocket connection totals are not a count of natural persons.",
-    "anonymousTelemetryConsentFooter": "No usage statistics are sent before you choose. You can change this choice later under Settings → System Settings.",
-    "anonymousTelemetryDecline": "Decline and disable",
-    "anonymousTelemetryAccept": "Agree and enable",
+    "anonymousTelemetryConsentTitle": "Help improve TX-5DR",
+    "anonymousTelemetryBannerSummary": "Share a small amount of runtime information so we can understand versions and environments and prioritize compatibility and stability improvements.",
+    "anonymousTelemetryBannerPrivacy": "Currently this includes only version, runtime environment, and runtime status—not callsigns, QSO records, frequencies, audio, or logs. You can turn it off at any time.",
+    "anonymousTelemetryCollectedTitle": "When enabled, TX-5DR sends:",
+    "anonymousTelemetryCollected": "the app version, runtime environment (system type, processor, and installation method), online status, uptime, and connection count. A random installation ID prevents duplicate counting.",
+    "anonymousTelemetryNeverCollectedTitle": "The statistics do not include:",
+    "anonymousTelemetryNeverCollected": "callsigns, QSO records, frequencies, audio, or log contents.",
+    "anonymousTelemetryMeasurementNote": "The totals represent installations, not individual people.",
+    "anonymousTelemetryDecline": "Don't share",
+    "anonymousTelemetryAccept": "Share statistics",
     "anonymousTelemetryConsentSaveFailed": "We could not save your choice. Please try again.",
     "loadFailed": "Failed to load configuration",
     "saveFailed": "Failed to save configuration",
