+++ packages/web/src/i18n/locales/ja/common.json
@@ -256,22 +256,20 @@
   "system": {
     "title": "システム設定",
     "description": "システムレベルの機能オプションを構成します。",
-    "anonymousTelemetryTitle": "匿名利用統計",
-    "anonymousTelemetryDesc": "同意いただいた場合、統計に参加するインストール数、利用状況、使用中のバージョン、同時オンラインインスタンス数の把握に役立てます。",
-    "anonymousTelemetryConsent": "プライバシーに関する選択がまだ完了していません。確認するまで利用統計は送信されません。",
+    "anonymousTelemetryConsent": "統計に参加するかまだ選択されていません。選択するまで統計データは送信されません。",
     "anonymousTelemetryUnavailable": "このビルドには統計サービスの接続先が設定されていないため、データは送信されません。",
     "anonymousTelemetryEnabled": "有効",
     "anonymousTelemetryDisabled": "無効",
-    "anonymousTelemetryConsentTitle": "匿名利用統計とプライバシーについて",
-    "anonymousTelemetryConsentIntro": "TX-5DR は、実際の利用状況を把握して今後のリリースを改善するため、最小限の匿名ランタイムデータを収集します。",
-    "anonymousTelemetryCollectedTitle": "収集する情報",
-    "anonymousTelemetryCollected": "ランダムに生成されたインストール識別子、アプリのバージョンとビルド識別子、リリースチャネル、実行形態、OS の種類、CPU アーキテクチャ、プロセスの開始・終了とオンライン状態、稼働時間、ハンドシェイク済み WebSocket 接続数です。サービスにはインストール識別子から一方向変換したキーのみを保存します。",
-    "anonymousTelemetryNeverCollectedTitle": "収集しない情報",
-    "anonymousTelemetryNeverCollected": "コールサイン、グリッドロケーター、QSO 記録、周波数、音声、無線機モデル、シリアル番号、ホスト名、ユーザー名、ファイルパス、URL、トークン、生の IP アドレス、User-Agent、ログ内容です。診断ログが自動的にアップロードされることもありません。",
-    "anonymousTelemetryMeasurementNote": "インストール数は統計に参加するインストール実体、アクティブ数とオンライン数は実行中のインスタンスを表します。WebSocket 接続数も個人の人数ではありません。",
-    "anonymousTelemetryConsentFooter": "選択するまで利用統計は送信されません。この設定は後から「設定 → システム設定」でいつでも変更できます。",
-    "anonymousTelemetryDecline": "同意せず統計を無効化",
-    "anonymousTelemetryAccept": "同意して統計を有効化",
+    "anonymousTelemetryConsentTitle": "TX-5DR の改善にご協力ください",
+    "anonymousTelemetryBannerSummary": "少量の実行情報を共有し、使用中のバージョンと環境を把握して、互換性と安定性の改善を優先できるようにします。",
+    "anonymousTelemetryBannerPrivacy": "現在含まれるのはバージョン、実行環境、実行状態のみで、コールサイン、QSO 記録、周波数、音声、ログは含まれません。設定からいつでも無効にできます。",
+    "anonymousTelemetryCollectedTitle": "有効にすると送信される情報：",
+    "anonymousTelemetryCollected": "アプリのバージョン、実行環境（システムの種類、プロセッサ、インストール方法）、オンライン状態、稼働時間、接続数です。重複を避けるため、インストールごとにランダムな ID を使用します。",
+    "anonymousTelemetryNeverCollectedTitle": "統計に含まれない情報：",
+    "anonymousTelemetryNeverCollected": "コールサイン、QSO 記録、周波数、音声、ログ内容です。",
+    "anonymousTelemetryMeasurementNote": "統計結果はインストール数を表し、実際の利用者数ではありません。",
+    "anonymousTelemetryDecline": "共有しない",
+    "anonymousTelemetryAccept": "統計を共有する",
     "anonymousTelemetryConsentSaveFailed": "選択を保存できませんでした。もう一度お試しください。",
     "loadFailed": "構成のロードに失敗しました",
     "saveFailed": "設定の保存に失敗しました",
