+++ packages/web/src/i18n/locales/zh/common.json
@@ -256,22 +256,20 @@
   "system": {
     "title": "系统设置",
     "description": "配置系统级别的功能选项。",
-    "anonymousTelemetryTitle": "匿名使用情况统计",
-    "anonymousTelemetryDesc": "经您同意后，帮助我们了解参与统计的安装量、使用量、活跃版本和同时在线实例数。",
-    "anonymousTelemetryConsent": "您尚未完成隐私选择；确认前不会发送任何统计数据。",
+    "anonymousTelemetryConsent": "您还没有选择是否参与。在此之前，不会发送统计数据。",
     "anonymousTelemetryUnavailable": "当前构建未配置统计服务地址，因此不会发送任何数据。",
     "anonymousTelemetryEnabled": "已开启",
     "anonymousTelemetryDisabled": "已关闭",
-    "anonymousTelemetryConsentTitle": "匿名使用统计与隐私说明",
-    "anonymousTelemetryConsentIntro": "TX-5DR 希望收集最少量的匿名运行数据，以了解软件的实际使用情况并改进后续版本。",
-    "anonymousTelemetryCollectedTitle": "会收集",
-    "anonymousTelemetryCollected": "随机生成的安装实例标识；应用版本与构建标识、发布渠道、运行形态、操作系统类别和 CPU 架构；程序启停与在线状态、运行时长，以及已完成握手的 WebSocket 连接数。服务端仅保存由安装标识转换出的不可逆键值。",
-    "anonymousTelemetryNeverCollectedTitle": "绝不收集",
-    "anonymousTelemetryNeverCollected": "呼号、网格坐标、QSO 记录、频率、音频、电台型号、设备序列号、主机名、用户名、文件路径、URL、令牌、原始 IP 地址、User-Agent 或日志内容，也不会自动上传诊断日志。",
-    "anonymousTelemetryMeasurementNote": "安装量表示参与统计的安装实例数；活跃量和同时在线量表示运行实例数；WebSocket 连接数也不等同于自然人数。",
-    "anonymousTelemetryConsentFooter": "作出选择前不会发送任何统计数据。您之后可以随时前往“设置 → 系统设置”修改此选择。",
-    "anonymousTelemetryDecline": "不同意并关闭统计",
-    "anonymousTelemetryAccept": "同意并启用统计",
+    "anonymousTelemetryConsentTitle": "帮助改进 TX-5DR",
+    "anonymousTelemetryBannerSummary": "分享少量运行信息，帮助我们了解版本和运行环境，优先改进兼容性与稳定性。",
+    "anonymousTelemetryBannerPrivacy": "当前仅包含版本、运行环境和运行状态，不包含呼号、通联记录、频率、音频或日志；可随时在设置中关闭。",
+    "anonymousTelemetryCollectedTitle": "开启后会发送：",
+    "anonymousTelemetryCollected": "软件版本、运行环境（系统类型、处理器和安装方式）、在线状态、运行时长和连接数。每个安装会使用一个随机编号，避免重复计算。",
+    "anonymousTelemetryNeverCollectedTitle": "统计中不包含：",
+    "anonymousTelemetryNeverCollected": "呼号、通联记录、频率、音频或日志内容。",
+    "anonymousTelemetryMeasurementNote": "统计结果代表安装实例，不代表实际用户人数。",
+    "anonymousTelemetryDecline": "不参与",
+    "anonymousTelemetryAccept": "参与统计",
     "anonymousTelemetryConsentSaveFailed": "无法保存您的选择，请重试。",
     "loadFailed": "加载配置失败",
     "saveFailed": "保存配置失败",
